Added doors mode check to panels
Made it so various (Panel) items check the the doors mode and only count as granting access in Panels mode. This is so that getting something like the Windmill Entry (Panel) doesn't look like Windmill access in Simple/Complex/Max Doors.
The Panels/Max-exclusive items can be left as-is because they're never visible in Simple/Complex Doors in the first place.