Fix image filename casing for Linux compatibility
Some PNGs didn't show up when running PopTracker on Linux because the
filenames had the wrong case. This renames these files to use lower-case
extensions (as specified in the JSON files) as well as changing the JSON
itself to use lower-case extensions in some cases.

The Doors*.png files were renamed to doors*.png, as this seems like the
intended casing.