Remove a snipe
Technically, the snipe is False if non-randomizeds are disabled (and not solvable). And it just so happens to work out that the alternative way of doing this snipe requires the exact same logic as this method (that is: no Symbols and just overworld access). So this can safely be deleted instead of awkwardly worked around